Toying and Notes
Updated some of the documentation for the qs_score
algorithm but I have yet to really write the algorithm.
Also, I played around a bit with the assert! macro,
although there's not much in the test file as of right
now. Anyways, just commiting this before I forget and it
ends up going in with some actually useful stuff.