Finished initial port of QuickSilver comparison
This is the first-pass attempt at porting over the quicksilver algorithm
for Rust. I'm still quite new to Rust so I'm sure there is lots of love
that could be given to this code to make it "real" or "proper" Rust code,
but I imagine it'll get there with time (if needed).